refactor: seperate routes and controllers, and fix some type errors
