+++ apps/backend/index.ts
@@ -13,7 +13,7 @@ import { authMiddleware } from "./middleware";
 import dotenv from "dotenv";
 
 import paymentRoutes from "./routes/payment.routes";
-import {router as webhookRouter} from './routes/webhook.routes';
+import webhookRouter from "./routes/webhook.routes";
 
 const IMAGE_GEN_CREDITS = 1;
 const TRAIN_MODEL_CREDITS = 20;
@@ -355,7 +355,7 @@ app.post("/fal-ai/webhook/image", async (req, res) => {
 });
 
 app.use("/payment", paymentRoutes);
-app.use("/api/webhook",webhookRouter );
+app.use("/api/webhook", webhookRouter);
 
 app.listen(PORT, () => {
   console.log(`Server is running on port ${PORT}`);
